🎨 (express-request-validator): Move request validator to shared package
Also renames error-handler to express-error-handler
