+++ apps/consent-api/package.json
@@ -21,7 +21,8 @@
 		"axios": "^1.5.0",
 		"body-parser": "^1.20.2",
 		"cors": "^2.8.5",
-		"error-handler": "workspace:^",
+		"express-error-handler": "workspace:^",
+		"express-request-validation": "workspace:^",
 		"express": "^4.18.2",
 		"logger": "workspace:^",
 		"types": "workspace:^",
