+++ apps/consent-das/package.json
@@ -1,39 +1,39 @@
 {
-  "name": "consent-das",
-  "version": "0.1.0",
-  "description": "Consent Data Access Service",
-  "exports": null,
-  "type": "module",
-  "engines": {
-    "node": ">=18"
-  },
-  "scripts": {
-    "build": "tsc",
-    "dev": "nodemon",
-    "start": "ts-node src/server.ts",
-    "migrate-dev": "pnpm prisma migrate dev",
-    "generate": "pnpm prisma generate",
-    "lint": "eslint .",
-    "seed": "ts-node ./prisma/seed/seed.ts"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@prisma/client": "^5.1.1",
-    "body-parser": "^1.20.2",
-    "error-handler": "workspace:^",
-    "express": "^4.18.2",
-    "logger": "workspace:^",
-    "nanoid": "^5.0.2",
-    "types": "workspace:^"
-  },
-  "devDependencies": {
-    "@types/body-parser": "^1.19.2",
-    "@types/express": "^4.17.17",
-    "dotenv": "^16.3.1",
-    "eslint-config-ohcrn": "workspace:^",
-    "nodemon": "^3.0.1",
-    "prisma": "^5.1.1"
-  }
+	"name": "consent-das",
+	"version": "0.1.0",
+	"description": "Consent Data Access Service",
+	"exports": null,
+	"type": "module",
+	"engines": {
+		"node": ">=18"
+	},
+	"scripts": {
+		"build": "tsc",
+		"dev": "nodemon",
+		"start": "ts-node src/server.ts",
+		"migrate-dev": "pnpm prisma migrate dev",
+		"generate": "pnpm prisma generate",
+		"lint": "eslint .",
+		"seed": "ts-node ./prisma/seed/seed.ts"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"@prisma/client": "^5.1.1",
+		"body-parser": "^1.20.2",
+		"express-error-handler": "workspace:^",
+		"express": "^4.18.2",
+		"logger": "workspace:^",
+		"nanoid": "^5.0.2",
+		"types": "workspace:^"
+	},
+	"devDependencies": {
+		"@types/body-parser": "^1.19.2",
+		"@types/express": "^4.17.17",
+		"dotenv": "^16.3.1",
+		"eslint-config-ohcrn": "workspace:^",
+		"nodemon": "^3.0.1",
+		"prisma": "^5.1.1"
+	}
 }
