+++ apps/pi-das/package.json
@@ -22,7 +22,7 @@
 	"dependencies": {
 		"@prisma/client": "^5.1.1",
 		"body-parser": "^1.20.2",
-		"error-handler": "workspace:^",
+		"express-error-handler": "workspace:^",
 		"express": "^4.18.2",
 		"logger": "workspace:^",
 		"nanoid": "^5.0.2",
