+++ packages/express-error-handler/package.json
@@ -1,0 +1,25 @@
+{
+	"name": "express-error-handler",
+	"version": "0.1.0",
+	"description": "OHCRN Platform Error Handler",
+	"exports": "./dist/index.js",
+	"types": "dist/index.d.ts",
+	"type": "module",
+	"engines": {
+		"node": ">=18"
+	},
+	"scripts": {
+		"build": "tsc"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"express": "^4.18.2",
+		"logger": "workspace:^",
+		"types": "workspace:^"
+	},
+	"devDependencies": {
+		"@types/express": "^4.17.17"
+	}
+}
