+++ packages/express-logger/reference-log-objects.md
@@ -1,0 +1,54 @@
+## Full Default Log Messages 
+
+### Request
+
+```json
+{
+  "request": {
+    "method": "GET",
+    "url_params": {},
+    "url": "/consent-questions?category=INFORMED_CONSENT",
+    "url_route": "N/A",
+    "query": { "category": "INFORMED_CONSENT" },
+    "headers": { },
+    "timestamp": "2023-11-19T17:00:23.414Z",
+    "timestamp_ms": 1700413223414,
+    "body": "N/A"
+  },
+  "millis-timestamp": 1700413223415,
+  "utc-timestamp": "2023-11-19T17:00:23.415Z",
+  "stage": "start"
+}
+```
+
+### Response
+
+```json
+{
+  "response": {
+    "status_code": 304,
+    "timestamp": "2023-11-19T17:00:23.421Z",
+    "timestamp_ms": 1700413223421,
+    "elapsed": 7,
+    "headers": {
+      "x-powered-by": "Express",
+      "etag": "W/\"21-NacCLjfDp1XGvWKo4cKckOSvxE4\""
+    },
+    "body": "N/A"
+  },
+  "request": {
+    "method": "GET",
+    "url_params": {},
+    "url": "/?category=INFORMED_CONSENT",
+    "url_route": "/consent-questions/",
+    "query": { "category": "INFORMED_CONSENT" },
+    "headers": { },
+    "timestamp": "2023-11-19T17:00:23.414Z",
+    "timestamp_ms": 1700413223414,
+    "body": "N/A"
+  },
+  "millis-timestamp": 1700413223421,
+  "utc-timestamp": "2023-11-19T17:00:23.421Z",
+  "stage": "end"
+}
+```
