+++ packages/express-request-validation/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "error-handler",
+	"name": "express-request-validation",
 	"version": "0.1.0",
 	"description": "OHCRN Platform Error Handler",
 	"exports": "./dist/index.js",
@@ -9,17 +9,20 @@
 		"node": ">=18"
 	},
 	"scripts": {
-		"build": "tsc"
+		"build": "tsc",
+		"test": "vitest",
+		"coverage": "vitest --coverage"
 	},
 	"keywords": [],
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
 		"express": "^4.18.2",
-		"logger": "workspace:^",
-		"types": "workspace:^"
+		"types": "workspace:^",
+		"zod": "^3.21.4"
 	},
 	"devDependencies": {
-		"@types/express": "^4.17.17"
+		"@types/express": "^4.17.17",
+		"@types/express-serve-static-core": "^4.17.41"
 	}
 }
