+++ packages/express-request-validation/tsconfig.json
@@ -1,0 +1,13 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"outDir": "dist",
+		"baseUrl": ".",
+		"declaration": true
+	},
+	"ts-node": {
+		"esm": true,
+		"logError": true
+	},
+	"include": ["src", "tests"]
+}
